+++ tests/Framework/Kernel.php
@@ -16,8 +16,10 @@ namespace Rekalogika\DomainEvent\Tests\Framework;
 use Doctrine\Bundle\DoctrineBundle\DoctrineBundle;
 use Rekalogika\DomainEvent\RekalogikaDomainEventBundle;
 use Symfony\Bundle\FrameworkBundle\FrameworkBundle;
+use Symfony\Component\Config\FileLocator;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Loader\PhpFileLoader;
 use Symfony\Component\HttpKernel\Kernel as HttpKernelKernel;
 
 class Kernel extends HttpKernelKernel
@@ -42,6 +44,16 @@ class Kernel extends HttpKernelKernel
         ];
     }
 
+    public function build(ContainerBuilder $container): void
+    {
+        $loader = new PhpFileLoader(
+            $container,
+            new FileLocator(__DIR__ . '/Resources/config')
+        );
+
+        $loader->load('services_test.php');
+    }
+
     public function getProjectDir(): string
     {
         return dirname(dirname(__DIR__));
@@ -54,7 +66,7 @@ class Kernel extends HttpKernelKernel
 
     public function registerContainerConfiguration(LoaderInterface $loader): void
     {
-        $loader->load($this->getConfigDir() . '*' . '.yaml', 'glob');
+        $loader->load($this->getConfigDir() . '/packages/*' . '.yaml', 'glob');
 
         $loader->load(function (ContainerBuilder $container) {
             $container->loadFromExtension('rekalogika_domain_event', $this->config);