+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * test: Overhaul tests
 * fix: Decorates the main Doctrine service IDs, not the aliases.
 * fix: Use `callable|array` typehint when decorating event dispatchers. See https://github.com/symfony/symfony/issues/48130
+* test: refactor tests
 
 ## 1.2.5